Run on port 8000 by default.
Port 80 is privileged. While the Dockerfile currently instructs the
container to run as root, in environments like our Kubernetes cluster we
do not want the service to run as root.

We can also theoretically drop to an unprivileged user here (for
instance using `USER nobody`), this is left out for now.

References: https://github.com/python-discord/kubernetes/pull/103